+++ state-manager.js
@@ -1880,7 +1880,7 @@ function getSettingsInternal(extensionSettings) {
          s.stockPrompts.combat.includes('BAB is +10') ||
          (s.stockPrompts.combat.includes('Att/def:') && !s.stockPrompts.combat.includes('(N attacks,')) ||
          (s.stockPrompts.combat.includes('Att/def:') && !s.stockPrompts.combat.includes('Spells:')) ||
-         (s.stockPrompts.combat.includes('Spells:') && !s.stockPrompts.combat.includes('Spell Atk')))) {
+         (s.stockPrompts.combat.includes('Spell Atk') && !s.stockPrompts.combat.includes('Spells: Cantrips:')))) {
         s.stockPrompts.combat = DEFAULT_STOCK_PROMPTS.combat;
     }